msm: thermal: simple: Introduce simple MSM thermal solution
When properly configured, this driver monitors temperatures from a
specified sensor and reduces the CPU's max frequency in response to
the temperature crossing specified thermal zones. The temperature sensor
use must be a specified ADC channel on a given VADC device, hence this
driver's dependency on the chipset being an MSM product.

This driver can only be configured via the device tree; it cannot be
configured at runtime. Configuration instructions can be found in the
accompanying documentation.

[yaro: Ported to 4.14.]
Co-authored-by: Yaroslav Furman <yaro330@gmail.com>

[deu: ported to 4.19]
Co-authored-by: deu <fawwazzuladhim700@gmail.com>

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Carlos Ayrton Lopez Arroyo <15030201@itcelaya.edu.mx>
Signed-off-by: kvsnr113 <kvsnrprojkt113@gmail.com>